signuplogin/requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
twilio==7.4.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
